pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ion_CSP"
dynamic = ["version"]
keywords = ["Ion_Crystal, Crystal_Structure_Prediction"]
authors = [
    {name = "Ze Yang", email = "yangze1995007@163.com"}
]
description = "Crystal Structure Design Software Based on Molecular/Ionic Configuration."
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "ase==3.23.0",
    "scipy==1.15.2",
    "deepmd-kit==3.0.1",
    "torch==2.5.0",
    "dpdispatcher==0.6.7",
    "numpy==1.26.4",
    "paramiko==3.5.1",
    "pyxtal==1.0.4",
    "phonopy==2.28.0",
    "rdkit==2024.3.3",
    "pytest==8.3.4",
]

[tool.setuptools.package-dir]
src = "src"

[tool.setuptools_scm]

[project.scripts]
ion-csp = "ion_CSP.task_manager:main"

[tool.pytest.ini_options]
addopts = [
    "--tb=short",
    "-v",
    "--cov=ion_CSP",
    "--cov-report=html",
    "--cov-branch",
]
filterwarnings = [
    "ignore::pytest.PytestAssertRewriteWarning",
]